Add support for glance multistore
Add parameters to support deploying multiple glance-api backends. The
existing GlanceBackend parameter specifies the first backend, and it's
the default backend when multiple backends are defined. Additional backends
are defined using a new GlanceMultistoreConfig parameter. The data is
a hash where the keys are the glance backend identifier and the data is
the glance settings for that backend. A new GlanceStoreDescription lets
the user to provide a description for each backend.

The configuration can specify any combination of supported backend types.
Multiple rbd backends can be specified, but cinder, file and swift
backends are limited to one each.

Change-Id: Ifc538250815f0ce0e534cae443037e143309ca56
Depends-On: I41ab9b3593bf3d078c5bbd1826df8308e3f5e7af
(cherry picked from commit 96937674a7524275b52d90d4b3a0f94c0e384937)