Remove stray conditional from aodh-evaluator tasks
In Ib4c59302ad5ad64f23419cd69ee9b2a80333924e a conditional
was moved incorrectly and left as a stray. Luckily it is
only a cosmetic issue, but it might be confusing when doing
any maintenance work so it's better to remove it.

Change-Id: Iac3cbf23437bb11ec47dd1a0f189babe0c5587d0
(cherry picked from commit edd10576ca33337b09070da204a88d189456a5e2)