Fix kill-script
This change squashes two changes that are being backported from the
tripleo-ansible repo.  They are to the same file, which was relocated to
tripleo-ansible between Train and Ussuri.

Change https://review.opendev.org/703123: In the kill-script there is a
string "Unknown action ${SIG} for ${$CT_NAME} ${CT_ID}" which results in
a "bad substitution" error, as there is no variable named with what the
contents of the CT_NAME environment variable contains.  Remove the
extraneous '$'.

Change https://review.opendev.org/703128: The reason bug #1860155 was
triggered was because the kill script did not have a stanza for handling
the signal that was passed in, which is signal 15.  Since signal 15 is
unhandled, keepalived processes will still stick around.  Add handling
for signal 15.

Change-Id: Ib47fc73b498b6366efa4ae5b16855bd45cb3ec91
Fixes-Bug: #1860155