Remove dracut-config-generic package
In LP#1830574, we introduce the dracut command to include necessary
file into initramfs. However, dracut in RHEL or CentOS doesn't include
sysctl.conf or other specific module related confs at an installation
of new kernel if dracut-config-generic package exists on the system.

We should remove the package to allow creating a host-specific initramfs
at an installation of new kernel.

Change-Id: I8e4c675925cdae292c25a1e86b904fd7f618633f
Closes-bug: #1857493
Co-Authored-By: Keigo Noha <knoha@redhat.com>