Remove dracut-config-generic package.
In LP#1830574, we introduce the dracut command to include necessary
file into initramfs. However, dracut in RHEL or CentOS doesn't include
sysctl.conf or other specific module related confs at an installation
of new kernel if dracut-config-generic package exists on the system.

We should remove the package to allow creating a host-specific initramfs
at an installation of new kernel.

Change-Id: Idbb69f4e05681f86933a7b57a506221cf6581536
Closes-bug: #1857493
Co-Authored-By: Keigo Noha <knoha@redhat.com>