Force facts cache refreshing after OS upgrade.
When upgrading from Rocky to Stein, an upgrade of the operating system is
performed. This upgrade from RHEL7 to RHEL8 implies the removal of the
default /usr/bin/python binary. As the facts cache is enabled, Ansible's
strategy does not consider to upgrade facts and therefore we try to run the
ansible playbook using the old python binary when running the upgrade.
This fails with the error: /usr/bin/python: No such file or directory.

This patch makes use of the setup task in combination with clear_facts
right after rebooting to upgrade the operating system, to ensure that
the facts are gathered and refreshed for the Overcloud node just upgraded.
This way, we make sure that we are using the right python binary.

Closes-Bug: #1856313
Change-Id: Ia1fa60c22e482ab14a509730cf93634772e077a7