Swap tasks to avoid non-existing variable error.
The upgrade workflow to Stein has a guard task that
checks that the --limit option is being used when
running the overcloud upgrade run command, as the
upgrade needs to be performed node by node due to
the operating system upgrade. However, if the --limit
option is not passed, the upgrade tasks fails in the
task right before the guard, as that task already
references the undefined variable. So, it is needed
to invert the order so we fail at will in the guard
task.

Closes-Bug: #1861663
(cherry picked from commit bca35e611bbc8f103e4893282c9b2c49aa4e1e48)

Change-Id: I1def295469875e648ad1c9d1e0de00b241acc3fd