Swap tasks to avoid non-existing variable error.
The upgrade workflow to Stein has a guard task that
checks that the --limit option is being used when
running the overcloud upgrade run command, as the
upgrade needs to be performed node by node due to
the operating system upgrade. However, if the --limit
option is not passed, the upgrade tasks fails in the
task right before the guard, as that task already
references the undefined variable. So, it is needed
to invert the order so we fail at will in the guard
task.

Change-Id: I9ffddcaa52314c615362969757c94ebdf01a3b6d
Closes-Bug: #1861663