[update] Ensure we get fresh hiera data before running update_steps.
We need to make sure that the hiera data are fresh before the update
step so that anyone using those data during those steps are seeing the
latest information from heat.

Factor out the hiera generation and include it in deployment and
update playbooks.

The double tasks definition in the deployment playbook seemed to be
redundant so It has been removed.

Change-Id: I6b6c676880ccc8cbed23af135e5865c222a8f1d0
Closes-Bug: #1861799
With some adjustment:
(cherry picked from commit 536230b323e5c3ac41e6ed990360317cdb5bc702)