Replace '' by [] when a bind mount isn't needed
To avoid empty volumes like:

{
  (...)
  "volumes": [
    "/etc/puppet:/etc/puppet:ro",
    (...)
    "",
    ""
  ],
}

Replace '' by [], so heat won't create an item in the list.
It helps to have idempotent containers, since podman_container module
will compare the list of volumes that is given in parameters (containing
the empty entries) vs the list of volumes actually in podman inspect.
Replacing to [] clears out empty volumes and makes these containers
idempotent when podman_container module is used to deploy containers.

Change-Id: I228b01009e7d9980bee5480778dbc88b9e226297