Create /var/lib/config-data if it doesn't exist.
When performing the Undercloud upgrade from OSP13
to OSP16, we start from an almost empty Undercloud
node which has been upgraded to RHEL8 via Leapp.
The /var/lib/config-data is being lost during the
upgrade procedure, so this task makes sure that the
folder exists before checking the selinux state.

Change-Id: I760a4e532e0c299efcf57cee68e8e8f93795ea29