Fail NetworkConfig task on timeout
Running os-net-config as async, with failed_when as false
results in undefined variable error if async task times out.
Instead of ignoring failure of task, check for the
presense of results of the command execution 'rc', if it
is not defined, then rest of the tasks are not useful.
Closes-Bug: #1862627

Change-Id: Ibbcde856ac69bf73a47086d95a52c3b1a0d10911
(cherry picked from commit f17e06a778e7663bcd88d6d83474d426f62ef6be)