[OVN] Split NeutronWorkers and NeutronMetadataWorkers for OVN
Prior to this patch, the OVN metadata reused the NeutronWorkers
configuration to set the number workers it should spawn. This is not
ideal for OVN metadata because it works differently than the OVS
metadata agent, for example:

1) The OVN metadata workers are deployed onto the compute nodes and not
   on the controllers/gateways as the OVS ones are. This means it's
   better distributed.

2) Even tho we could have multiple workers, the OVN SB IDL in the OVN
   metadata agents acquires a lock while listening to OVSDB events.
   Therefore, even tho we could have dozens of agents only one of them
   is listening to the events.

3) Instantiating IDLs means maintaining in-memory replicas of the
   OVSDB (or, the tables of that DB that is being watched). This puts a
   burden on OVSDBs to maintain all these replicas updated, reducing
   this number workers also reduces the burden on the OVSDBs.

With this patch, we are adding a new configuration called
NeutronMetadataWorkers so that we can set different values for the API
workers and the metadata workers in a OVN deployment.

Change-Id: I5f0151d70977b3fdbf04c41ffb224f364a027e5c
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>