Configure Undercloud hostname in the overcloud during upgrade.
Due to IPv6, the undercloud's container registry had to change
from an IP address in the full hostname [0]. This impacts into
the upgrade to Train as the overcloud nodes do not contain the
Undercloud's host in /etc/hosts as well as it's missing in the
registries.conf.

This patch adds two new upgrade tasks to handle:

1. container-image-prepare: Ensure that /etc/hosts file contains
an entry for the Undercloud's ctrlplane hostname. This task makes
use of the global_vars ansible parameter undercloud_hosts_entry
which contains a list of the undercloud's ctrlplan hostnames (calling
getent hosts underneath).

2. podman-baremetal-ansible: There is already an upgrade task which
takes care of reconfiguring podman during the upgrade, so this patch
simply sets up the right container unsecure registries and passes it
into the reconfiguring task. It also changes the (step | int) into
step|int as the upgrade tasks require step|int condition to decide in
which step_X playbook fall, otherwise the task will appear in every
step_X playbook.

[0] - Iac6efde9dd283906274d95c3a239b4b882ec052e

Depends-On: https://review.opendev.org/708128
Closes-Bug: #1863598
Change-Id: Ifadc797f33d759eed38c9d9274fa588b6dd19488