Override nova's glance endpoint only when necessary
nova::glance_api_servers is deprecated because the corresponding nova
parameter is deprecated. There is a new nova::glance_endpoint_override,
but it should be set only in situations where the intent is to *not*
use ksa to discover glance's endpoint. For example, in a DCN/Edge
deployment, we override the endpoint to force nova to use the glance
service running at the edge site.

Change-Id: I42af3e39da76ae94ca7bbf2797f776c28a75f7e7
Depends-On: Ib7fac4f37ef02d8f577abc98e4cc78b750caba54