tripleo_container_manage: set tripleo_container_manage_check_puppet_config
Setting tripleo_container_manage_check_puppet_config to true will allow
us to get up-to-date config hashes in the container startup config right
before we decide if a container needs to be restarted or not.
It helps to have idempotent ansible runs and we avoid the restart of
containers when not needed.

Depends-On: I16b2972cdf79cd6ac927925607197ec2a969a28b
Change-Id: I88b29a13d8e7c149f52e41d9fe8aa1f5dda2ecd9