tripleo_container_manage: set tripleo_container_manage_check_puppet_config
Setting tripleo_container_manage_check_puppet_config to true will allow
us to get up-to-date config hashes in the container startup config right
before we decide if a container needs to be restarted or not.
It helps to have idempotent ansible runs and we avoid the restart of
containers when not needed.

Change-Id: I88b29a13d8e7c149f52e41d9fe8aa1f5dda2ecd9
(cherry picked from commit ddb47c23a66f4eee61343fd69ba20e3230fb3c77)