Remove BlockStorageCinderVolume service from certain DCN roles
The ScaleOut roles should accompany the DistributedCompute roles
which are sufficient to provide the BlockStorageCinderVolume
service.

The DistributedCompute role supports usecases without persistent
storage via Cinder while the DistributedComputeHCI supports
usecases with persistent storage via Cinder. For those usecases
we want the BlockStorageCinderVolume service to be used by
DistributedComputeHCI with Ceph but not without Ceph as that
is presently the only Cinder backend supporting active/active.

Change-Id: I8588919cecc2be06447eba2b53b79d8d7cfc6a9e
Fixes-Bug: #1863799