Remove support for ceph-ansible fetch directory
The state information in the ceph-ansible fetch directory
is no longer necessary to replace a Ceph monitor as of
ceph-ansible 4.0.14-1. Stein, Train, Ussuri and newer
should be using this version of ceph-ansible or newer.
TripleO does not need to do extra work to maintain the
fetch directory.

Add SwiftFetchDirGetTempurl, SwiftFetchDirPutTempurl, and
LocalCephAnsibleFetchDirectoryBackup to the deprecated list
of parameters and modify code to not use these parameters.
Remove call to tripleo-ansible/tripleo_ceph_fetch_dir role.

Update tripleo-common/workbooks/plan_management.yaml to not
create SwiftFetchDirGetTempurl and SwiftFetchDirPutTempurl
during deployment and update via depends-on patch.

Depends-On: I167449e4d7e38f0f5fd74251a3bd8f5b0b31451b
Change-Id: I3f8c2571a6dd35029d71dd42df637a962d9b3235
Closes-bug: #1863809