Use a jinja block to reduce str_replace repetition
It would appear that we use the same str_replace params most
of the time, and there's no harm in using aliases where only
some of them are used. This way we reduce the maintenance
burden because we only need to maintain it in one place.

Change-Id: Ib034405a15ade9e9fb234a9875ebbe922abfdfc6
(cherry picked from commit dda2030a6a8ed7062583a6e236059fc0b7d29f22)