Use a common playbook import for common plays
It would appear that we use many of the same plays in several
of the playbooks. In this patch we extract these into a common
playbook file so that we only need to repeat a single import
in each of the playbooks. This way we reduce the maintenance
burden because we only need to maintain it in one file.

Change-Id: Ib00e8aa9f7d06517290543a8aaf8a2527969bd3c