Add always tags for hieradata render in external upgrade.
When executing the external-upgrade command, if some of the
external_upgrade_tasks make use of the all_nodes vars (*_short_node_names
for example) then the command fails as these variables weren't
rendered. This only happens when we run the external-upgrade command
passing --tags, as the facts tags are being skipped.

This patch adds the always tags into those tasks which miss it for
external_upgrade_steps_playbook.yaml.

Change-Id: I0913df554b75d25215d5ae6547b939ad0356df3c