[OVN] SRIOV with native OVN DHCP server
In order to make SRIOV work for the OVN driver the concept of "external"
ports has been introduced (see depends-on). These ports lives on a
different host (gateway nodes) and are able to reply to arp requests
on behalf of VM port. In the SRIOV case, the SRIOV port is bypassed in
the hypervidor so the OVN driver creates an external port it which will
reply to the DHCP packets.

This patch is creating two new roles to work with the mechanism
described above:

* ControllerSriov: Same as the normal Controller role but with the OVN
Metadata agent deployed.

* NetworkerSriov: Same as the normal Networker role but with the OVN
Metadata agent deployed.

The patch also removes the Neutron DHCP agent from the
neutron-ovn-sriov.yaml environment file since no longer needed.

Depends-On: https://review.opendev.org/703376
Change-Id: I5ef3d6543785b677ea333803aaa23bd34abdd671
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>