Update the number of keystone workers
We recently patched puppet-keystone and puppet-openstacklib to set
KeystoneWorkers to a reasonable default given the consolidation of two
keystone processes into a single keystone process (keystone-main and
keystone-admin):

  https://review.opendev.org/#/c/702031/13
  https://review.opendev.org/#/c/705041/3

This patch updates THT to ensure we're using the new variable that's
dedicated to keystone's process count.

This ensures there isn't a performance regression due to reduced process
counts when upgrading to newer OpenStack versions because the keystone
applications were consolidated in puppet-keystone.

Change-Id: I47506181c38b9363b9c82c536e7c9d7d765f093f