[TRAIN-ONLY] Add systemd-wrapper update tasks
We backported the conversion for the side car containers wrapper tasks
from puppet to ansible in https://review.opendev.org/#/c/706381/.
However during a minor update, the deploy step tasks would be run after
the moutnts have been updated. We need to duplicate the deploy step
tasks for the wrapper creation so the mounts exist during the 'overcloud
update run process' as this runs before the 'overcloud update
converge'.

Change-Id: Ia898cb8b1c888aca45ba58ab99e61885a1da4f4e
Closes-Bug: #1866922