Use exists filter instead of stat where possible
When checking for the existance of a file on the host where
Ansible is being executed, using the stat module with localhost
delegation is rather heavy-handed. We can instead just make use
of the 'exists' test.

This should improve execution time just a little bit and reduces
the number of tasks for us to maintain. We also remove the
repetition of the task file path by using a variable.

Change-Id: I8b278ca83b2afb07575dbae2496ec265c3a06473
(cherry picked from commit 9666a7e6456b882e206436af085db0f6e97d3dab)