Use jinja raw tag instead of quoted concatenation
Many of the lines are difficult to grasp due to the crazy
quotation and concatenation implemented to get the desired
result from generating playbooks via a jinja template.

We can make it easier to read and easier to understand by
using the jinja raw tag instead. This eases the maintenace
burden on us all and helps us sleep better at night.

Change-Id: I82c4de4a63817707a2b0ed0ced827be37c0d0463