Skip operating system upgrade tasks via UpgradeLeappEnabled param.
The operating system from RHEL 7 to RHEL 8 was required during the upgrade
from Rocky to Stein, however it isn't anymore for the upgrade from Stein to
Train. But we can't get rid of these tasks as they will be required for the
three releases jump from Queens to Train.

The solution has been making use of an existing heat paramter
UpgradeLeappEnabled which will be set when an Operating System is required.
Before, this parameter defaulted to true, but from now on it
defaults to false and will be set to true during the prepare step.

Change-Id: I7ac0c74726f7bbeb773d54f6909c5f647717f79a