Use multipath in nova when multipathd is enabled
Currently we have environments/multipathd.yaml to deploy multipathd
service, but we still need to set NovaLibvirtVolumeUseMultipath to
make instances use multipath for volume connections.

This patch makes the environment file to also enable multipath usage
in nova, as we expect that user who use the environment want to
use multipath for volume connections.

Change-Id: I3ca55c01c4b03e990470c55b835095a71f3cad62