Revert "Stop using swift temp url for config transport"
This switch seems to be creating issues with upgrades, where a number of
software deployments are deleted concurrently while updating the config
transport for the server. Switching the config transport does not work
with convergence heat and should be fixed in heat. We can revert this
now, as we still use swift for other stuff in the undercloud. Can be
changed once the issue is fixed in heat.

It also reverts the following dependant commit.

Revert "Cleanup SoftwareConfigTransport"

This reverts commit (1821c01846a20da331959ff49fe8536f1e1bf86a and
3ea9dd4040686b7c2ec82f7e7e467e0b3c3bd2a7)

Closes-Bug: #1869335
Change-Id: I835c8be3eecce91f8a370d036bf1085bc445e01d
(cherry picked from commit 6ff119ddacdeb55d1637ad8c6557225732c31d22)