Switch scenario001-multinode-containers to LVM-based bluestore
scenario001-multinode was still using a loopback device directly
which is no longer supported as per the move to ceph-volume which
was documented in RH BZ 1666822. scenario00{1,4}-standalone were
updated to deal with this by passing precreated lvm_volumes to
CephAnsibleDisksConfig but scenario001-multinode was not. This
change brings it back up to date for deploying bluestore with
ceph-volume.

Change-Id: I1f0420b8d48d9a2e8327c78daf48449537e3b842