Workaround for cinder A/A and etcd with TLS-everywhere
This patch implements a workaround to allow cinder to run in active/
active mode with internal TLS enabled. Cinder uses etcd for its
distributed lock manager, and the LP bug documents several problems
when the deployment enables TLS on the internal API network.

Until a full solution is available, this workaround allows cinder and
etcd to work without TLS. The full solution is complicated, and affects
components outside of tripleo.

Change-Id: Iec0d02f8f51067098dd58beb4fe57a7fd5ab5651
Related-Bug: #1869955