Create DNS entries in IPA for openstack services
This adds forward and reverse DNS records for all services.

Without the reverse DNS entries, certificates cannot be created
for etcd which requires a subject alternative name (SAN) entry
to be included. This is used by cinder for A/A support.

The task is delegated to tripleo-ipa. It uses the host_entry
which is also used for managing the entries in /etc/hosts.

Change-Id: I41681f90f70fa0dffe4abbe3d6d5c48015589f66
(cherry picked from commit fe3f38d3a9a6d8795dd7109930ce433dc4488df6)
Related-bug: #1869955
Related: rhbz#1804079