Properly place undercloud hosts record upgrade task
Undercloud record for hosts has to be in all the nodes that fetch
images from it. Proper location for this code is where we set up
podman. If it's only in ContainerImagesPrepare it runs on the
Controller role only.

Closes-Bug: #1871243

Change-Id: Id081f049bca6c961ad2ff649ac357cea8d552739
(cherry picked from commit fff8cf73c6a4d0e311b8b02036e7df928b998339)