Exclude /etc/puppet from config generation
When we run the container puppet process, we nuke /etc/puppet in the
container and sync it in from the system. Since this is not part of the
actual configuration, we're getting changes to the hieradata which is
causing unnecessary container restarts.

Change-Id: I9e770a2e3342fbd9617ae4a59c7e617f84e9233d
Related-Bug: #1871734
(cherry picked from commit f8eb0a8fa04d96660a6f03e4187a1f246f0694f0)