Add hook to run RHOSP policies enforcement.
This add an import of the tripleo_redhat_enforce role[1] at the
beginning of undercloud upgrade tasks (which are used for update and
upgrade) and overcloud update.

The code upstream won't do anything even with SkipRhelEnforcement set
to 'false' as the role itself only activate the checks for RedHat
distribution. This variable is only there to support use case where we
deploy on Red Hat and for /some/ reason, the checks shouldn't
enabled (mainly CI).

[1] in tripleo-ansible

Depends-On: https://review.opendev.org/721292
Change-Id: I076bebf6bbd5f45d5ecb372a3b83e8cc279946b8