Set Neutron's l3_ha flag to True in standalone ML2/OVS job
Even if this is standalone job, when l3_ha config option is set
to True in Neutron server, all routers will be created as HA
routers.
So for each such router created during the tests things like
keepalived, neutron-keepalived-state-change-monitor will be spawned
and also HA network will be created for tenants used by Tempest.
This will give us better coverage of things which are most commonly
used in Neutron.

Depends-On: https://review.opendev.org/721799

Change-Id: I44ed60e88723e58ff1bb7bd0559540a596442f5a