Enable adding packages into Leapp's to_remove/to_install files.
Leapp enables removal and installing packages during the leapp upgrade, so
that the change takes effect during the machines reboot. This patch adds a
mechanism to configure that list of packages via two new heat parameters:
UpgradeLeappToInstall and UpgradeLeappToRemove.

Also, as part of the upgrade from Queens to Train, there is a big jump in the
openvswitch versions, having ovs2.11 in Queens and ovs2.13 in Train, which are
mutually exclusive. This package adds these two pacakges into the to_remove/
to_install mechanism so that the openvswitch package upgrade takes place
during Leapps reboot to upgrade the Operating System.

Change-Id: I2d07f19da68dd17b74189f273e75af3fe1d7a955