Fix validation script
This change adds support for special routes on the host. Without this
the validation fails if any routes are present that don't have an IP
address in the first column, like blackhole, prohibit, throw, etc.
Instead of piping a bunch of `grep -v`, the regex filters all cases by
capturing all lines that start with an IP address. The regex doesn't
capture only valid IP addresses, but it should do just fine here.

Change-Id: I933b33ccf22fb42695be6ec9a5e62506a1b355fd