Update Nova VNC puppet variables
We recently introduced new variables to puppet-nova that expose nova VNC
settings for SSL ciphers:

  I3a1262f70f6a801db276701a39ebb01f40025192

However, when we updated tht we used the actual names of the nova
settings instead of the names of the puppet variables:

  Ida03a0aa54ca15b343339d92abb9c105ead8b0b6

This patch changes tht to use the puppet variables instead of the nova
configuration names. Without this patch, setting the heat variables has
no affect on tls settings in nova.conf.

Change-Id: Iacc25c694f2e5491e971b9cef4e984602be25c04