Use tripleo linear when not using tripleo free
In order to handle percentage failures for tripleo classes we will need
to use a custom strategy to handle the failures. Let's explicitly define
our expected strategies in the playbook so it'll default to linear (the
ansible default) on the host when ansible is invoked but our deployment
tasks will be run as we want them.

Depends-On: https://review.opendev.org/#/c/763242/
Change-Id: Ifaddeccfbb82e03815311ee0b76eb2e2eae282a7